HACK: scripts: Makefile: disable size_append
techpack/audio is somehow triggering size_append
which is time-consuming, resource-hungry
and utterly meaningless during incremental builds.

Likely there are BUGs in Kbuild but so far I have
not been able to find them.

As we are not using size_append for this kernel,
let's just disable size_append before a proper
solution can be found.

Signed-off-by: Jesse Chan <jc@linux.com>